+++ pytest_factoryboy/__init__.py
@@ -1,6 +1,6 @@
 """pytest-factoryboy public API."""
 
-__version__ = '1.0.2'
+__version__ = '1.0.3'
 
 try:
     from .fixture import register